requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
scikit-learn
numpy
pandas
requests
pyyaml
streamlit
